refactor: 피드백을 반영했다.
